Fonctionnalité groupe Privé fini ,mise a niveau suppression participant ,mise a niveau bdd et gestion sécurité pour les groupes.

+++ src/Controller/GroupePriveController.php
@@ -6,9 +6,12 @@ use App\Entity\GroupePrive;
 use App\Entity\InscriptionGroupePrive;
 use App\Form\GroupePriveType;
 use App\Repository\GroupePriveRepository;
+use App\Repository\InscriptionGroupePriveRepository;
+use App\Repository\InscriptionsRepository;
 use App\Repository\ParticipantRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Bundle\SecurityBundle\Security;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
@@ -17,21 +20,42 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
-#[Route('/groupe/prive')]
+#[Route('/groupe_prive')]
 class GroupePriveController extends AbstractController
 {
     #[Route('/', name: 'app_groupe_prive_index', methods: ['GET'])]
-    public function index(GroupePriveRepository $groupePriveRepository): Response
+    public function index(Security $security, GroupePriveRepository $groupePriveRepository, InscriptionGroupePriveRepository $inscriptionGroupePriveRepository): Response
     {
+        // Récupérer l'utilisateur connecté
+        $user = $security->getUser();
+
+        // Récupérer les groupes appartenant à l'utilisateur connecté
+        $ownedGroupePrives = $groupePriveRepository->findBy(['owner' => $user]);
+
+        // Récupérer les groupes où l'utilisateur est inscrit
+        $inscriptions = $inscriptionGroupePriveRepository->findBy(['noParticipant' => $user]);
+
+        // Extraire les groupes des inscriptions
+        $subscribedGroupePrives = [];
+        foreach ($inscriptions as $inscription) {
+            $subscribedGroupePrives[] = $inscription->getNoGroupe();
+        }
+
+        // Fusionner les deux listes de groupes et enlever les doublons si nécessaire
+        $allGroupePrives = array_merge($ownedGroupePrives, $subscribedGroupePrives);
+        $allGroupePrives = array_unique($allGroupePrives, SORT_REGULAR);
+
         return $this->render('groupe_prive/index.html.twig', [
-            'groupe_prives' => $groupePriveRepository->findAll(),
+            'groupe_prives' => $allGroupePrives,
         ]);
     }
+
 
     #[Route('/new', name: 'app_groupe_prive_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $groupePrive = new GroupePrive();
+        $groupePrive->setOwner($this->getUser());
         $form = $this->createForm(GroupePriveType::class, $groupePrive);
         $form->handleRequest($request);
 
@@ -49,16 +73,50 @@ class GroupePriveController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_groupe_prive_show', methods: ['GET'])]
-    public function show(GroupePrive $groupePrive): Response
+    public function show(GroupePrive $groupePrive, ParticipantRepository $participantRepository,
+                         InscriptionGroupePriveRepository $inscriptionGroupePriveRepository): Response
     {
+        // Vérifier que l'utilisateur connecté est le propriétaire du groupe
+        if ($groupePrive->getOwner() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Access Denied');
+        }
+        // Récupérer tous les participants
+        $participants = $participantRepository->findAll();
+
+        // Créer un tableau de choix pour le formulaire
+        $lesInscrits = [];
+        foreach ($participants as $participant) {
+            // Vérifier si le participant est déjà inscrit dans ce groupe privé
+            $isAlreadyInscribed = $inscriptionGroupePriveRepository->findOneBy([
+                'noParticipant' => $participant->getId(),
+                'noGroupe' => $groupePrive->getId()
+            ]);
+
+            // Si le participant est inscrit, on l'ajoute
+            if ($isAlreadyInscribed) {
+                $lesInscrits[$participant->getPseudo()] = $participant;
+            }
+        }
         return $this->render('groupe_prive/show.html.twig', [
             'groupe_prive' => $groupePrive,
+            'lesInscrits' => $lesInscrits,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_groupe_prive_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, GroupePrive $groupePrive, EntityManagerInterface $entityManager): Response
+    public function edit(
+        Request $request,
+        GroupePrive $groupePrive,
+        EntityManagerInterface $entityManager,
+        ParticipantRepository $participantRepository,
+        InscriptionGroupePriveRepository $inscriptionGroupePriveRepository,
+    ): Response
     {
+        // Vérifier que l'utilisateur connecté est le propriétaire du groupe
+        if ($groupePrive->getOwner() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Access Denied');
+        }
+
         $form = $this->createForm(GroupePriveType::class, $groupePrive);
         $form->handleRequest($request);
 
@@ -71,12 +129,17 @@ class GroupePriveController extends AbstractController
         return $this->render('groupe_prive/edit.html.twig', [
             'groupe_prive' => $groupePrive,
             'form' => $form,
+
         ]);
     }
 
     #[Route('/{id}', name: 'app_groupe_prive_delete', methods: ['POST'])]
     public function delete(Request $request, GroupePrive $groupePrive, EntityManagerInterface $entityManager): Response
     {
+        // Vérifier que l'utilisateur connecté est le propriétaire du groupe
+        if ($groupePrive->getOwner() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Access Denied');
+        }
         if ($this->isCsrfTokenValid('delete'.$groupePrive->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($groupePrive);
             $entityManager->flush();
@@ -87,15 +150,35 @@ class GroupePriveController extends AbstractController
 
 
     #[Route('/addParticipants/{id}', name: 'app_groupe_prive_addParticipants')]
-    public function addParticipants(Request $request, GroupePrive $groupePrive, EntityManagerInterface $entityManager, ParticipantRepository $participantRepository, FormFactoryInterface $formFactory): Response
+    public function addParticipants(
+        Request $request,
+        GroupePrive $groupePrive,
+        EntityManagerInterface $entityManager,
+        ParticipantRepository $participantRepository,
+        FormFactoryInterface $formFactory,
+        InscriptionGroupePriveRepository $inscriptionGroupePriveRepository
+    ): Response
     {
+        // Vérifier que l'utilisateur connecté est le propriétaire du groupe
+        if ($groupePrive->getOwner() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Access Denied');
+        }
         // Récupérer tous les participants
         $participants = $participantRepository->findAll();
 
         // Créer un tableau de choix pour le formulaire
         $choices = [];
         foreach ($participants as $participant) {
-            $choices[$participant->getPseudo()] = $participant->getId(); // Utiliser l'ID comme valeur
+            // Vérifier si le participant est déjà inscrit dans ce groupe privé
+            $isAlreadyInscribed = $inscriptionGroupePriveRepository->findOneBy([
+                'noParticipant' => $participant->getId(),
+                'noGroupe' => $groupePrive->getId()
+            ]);
+
+            // Si le participant n'est pas encore inscrit, on l'ajoute aux choix
+            if (!$isAlreadyInscribed) {
+                $choices[$participant->getPseudo()] = $participant->getId(); // Utiliser l'ID comme valeur
+            }
         }
 
         // Créer le formulaire
@@ -142,4 +225,85 @@ class GroupePriveController extends AbstractController
             'form' => $form->createView(),
         ]);
     }
+    #[Route('/deleteParticipants/{id}', name: 'app_groupe_prive_deleteParticipants')]
+    public function deleteParticipants(
+        Request $request,
+        GroupePrive $groupePrive,
+        EntityManagerInterface $entityManager,
+        ParticipantRepository $participantRepository,
+        FormFactoryInterface $formFactory,
+        InscriptionGroupePriveRepository $inscriptionGroupePriveRepository
+    ): Response
+    {
+        // Vérifier que l'utilisateur connecté est le propriétaire du groupe
+        if ($groupePrive->getOwner() !== $this->getUser()) {
+            throw $this->createAccessDeniedException('Access Denied');
+        }
+        // Récupérer tous les participants
+        $participants = $participantRepository->findAll();
+
+        // Créer un tableau de choix pour le formulaire
+        $choices = [];
+        foreach ($participants as $participant) {
+            // Vérifier si le participant est déjà inscrit dans ce groupe privé
+            $isAlreadyInscribed = $inscriptionGroupePriveRepository->findOneBy([
+                'noParticipant' => $participant->getId(),
+                'noGroupe' => $groupePrive->getId()
+            ]);
+
+            // Si le participant n'est pas encore inscrit, on l'ajoute aux choix
+            if ($isAlreadyInscribed) {
+                $choices[$participant->getPseudo()] = $participant->getId(); // Utiliser l'ID comme valeur
+            }
+        }
+
+        // Créer le formulaire
+        $form = $formFactory->createBuilder(FormType::class)
+            ->add('participants', ChoiceType::class, [
+                'choices' => $choices,
+                'expanded' => true,  // Utilisation des cases à cocher
+                'multiple' => true,  // Permettre la sélection multiple
+                'label' => 'Select Participants',
+            ])
+            ->add('submit', SubmitType::class, ['label' => 'Delete Participants'])
+            ->getForm();
+
+        $form->handleRequest($request);
+
+        // Si le formulaire est soumis et valide
+        if ($form->isSubmitted() && $form->isValid()) {
+            // Récupérer les IDs des participants sélectionnés
+            $selectedParticipants = $form->get('participants')->getData();
+
+            // Créer les inscriptions pour les participants sélectionnés
+            foreach ($selectedParticipants as $participantId) {
+                $participant = $participantRepository->find($participantId);
+                if ($participant) {
+                    // Trouver l'inscription existante pour ce participant et ce groupe privé
+                    $inscription = $inscriptionGroupePriveRepository->findOneBy([
+                        'noGroupe' => $groupePrive,
+                        'noParticipant' => $participant
+                    ]);
+
+                    // Si une inscription existe, la supprimer
+                    if ($inscription) {
+                        $entityManager->remove($inscription);
+                    }
+                }
+            }
+
+            // Sauvegarder les nouvelles inscriptions en base de données
+            $entityManager->flush();
+
+            // Rediriger après l'inscription
+            return $this->redirectToRoute('app_groupe_prive_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        // Rendre le formulaire dans le template
+        return $this->render('groupe_prive/delete_participant.html.twig', [
+            'groupe_prive' => $groupePrive,
+            'form' => $form->createView(),
+        ]);
+    }
+
 }
